+++ sample/sample6-custom-decorator/IsLongerThan.ts
@@ -6,11 +6,17 @@ import {ValidationArguments} from "../../src/validation/ValidationArguments";
 
 export function IsLongerThan(property: string, validationOptions?: ValidationOptions) {
     return function (object: Object, propertyName: string) {
-        registerDecorator(object, propertyName, validationOptions, [property], IsLongerThanConstraint);
+        registerDecorator({
+            target: object.constructor,
+            propertyName: propertyName,
+            options: validationOptions,
+            constraints: [property],
+            validator: IsLongerThanConstraint
+        });
     };
 }
 
-@ValidatorConstraint("is_longer_than")
+@ValidatorConstraint({ name: "isLongerThan" })
 export class IsLongerThanConstraint implements ValidatorConstraintInterface {
     
     validate(value: any, args: ValidationArguments) {