+++ sample/sample6-custom-decorator/IsUserAlreadyExist.ts
@@ -4,15 +4,23 @@ import {ValidationArguments} from "../../src/validation/ValidationArguments";
 
 export function IsUserAlreadyExist(validationOptions?: ValidationOptions) {
     return function (object: Object, propertyName: string) {
-        const constraints: any[] = [/*constraints your decorator can have*/];
-        registerDecorator(object, propertyName, validationOptions, constraints, "user_exists", (userName, args) => {
-            return new Promise(ok => {
-                if (userName !== "admin" && userName !== "user") {
-                    ok(true);
-                } else {
-                    ok(false);
+        registerDecorator({
+            name: "isUserAlreadyExist",
+            async: true,
+            target: object.constructor,
+            propertyName: propertyName,
+            options: validationOptions,
+            validator: {
+                validate(value: any, args: ValidationArguments) {
+                    return new Promise(ok => {
+                        if (value !== "admin" && value !== "user") {
+                            ok(true);
+                        } else {
+                            ok(false);
+                        }
+                    });
                 }
-            });
+            }
         });
     };
 }